Renamed jni_writer_data_sink header.
