+++ java/src/main/native/src/csv_chunked_writer.hpp
@@ -19,7 +19,7 @@
 
 #include <cudf/io/csv.hpp>
 
-#include "jni_writer_data_sink.hpp"
+#include "writer_data_sink.hpp"
 
 namespace cudf::jni::io {
