+++ java/src/main/native/src/writer_data_sink.hpp
